+++ pkg/http-tunnel/server/registry.go
